Create 001150 Rachel Pollack.md
Pollack was the subject of a <a href="https://www.fifthestate.org/archive/380-spring-2009/rachel-pollack-willing-change-everything/">profile article</a> by Cara Hoffman, in _Fifth Estate_ #380. She notes that:

"Her work is often concerned with what is below the surface of everyday matters, what happens when the life of the mind and the concerns of the material world converge.

"This makes Pollack’s work not only transcendent, but often very funny. As in _Unquenchable Fire_, where God appears in the form of a chocolate chip cookie salesman on Seventh Avenue, and is then perceived to be out of touch with the experiences of mortals, and therefore insane."